requirements.txt
torchvision==0.9.1
torch==1.8.1
Flask_Bootstrap==3.3.7.1
Flask_MySQLdb==0.2.0
Flask==1.1.2
Pillow==8.2.0
PyYAML==5.4.1
gunicorn==20.1.0